requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
python-multipart
requests
